move CompositeBar into workbench/parts/compositebar
