+++ src/vs/workbench/browser/parts/compositebar/compositeBarActions.ts
